moved _convert_data_dataframe to _base.py
